Add Report0058.py
This report handles the following data criteria:

'Er zijn geen assets die het doel zijn van twee of meer Voedt relaties.'
