--- Reports/Report0049.py
@@ -1,22 +1,0 @@
-from DQReport import DQReport
-
-
-class Report0020:
-    def __init__(self):
-        self.report = None
-
-    def init_report(self):
-        self.report = DQReport(name='report0049',
-                               title='Er zijn geen assets die zichzelf direct of indirect voeden (geen lussen in voeding).',
-                               spreadsheet_id='',
-                               datasource='Neo4J',
-                               persistent_column='D')
-
-        self.report.result_query = """
-            MATCH p=(x:Asset {isActief: True})-[:Voedt*]->(x:Asset {isActief: True})
-            WITH x, reduce(path_loop = [], n IN nodes(p) | path_loop + [[n.uuid, n.typeURI]]) as path_loop
-            RETURN x.uuid AS uuid, x.naampad AS naampad, x.typeURI AS typeURI, path_loop
-        """
-
-    def run_report(self, sender):
-        self.report.run_report(sender=sender)
